chore: feedback PR, clarify name const + deplace const
