+++ src/modules/sessions/utils/session.utils.ts
@@ -1,17 +1,6 @@
-import type { SessionFormType, SessionPayload } from "../types/session-form.type";
+import type { SessionFormType, SessionPayload, SessionFormState } from "../types/session-form.type";
 
-export type SessionFormState = {
-  title: string;
-  description: string;
-  startDate: string;
-  startTime: string;
-  endTime: string;
-  linkedSkill: string;
-  linkedQuest: string;
-  color: string;
-};
-
-export const INITIAL_FORM: SessionFormState = {
+export const INITIAL_SESSION_FORM: SessionFormState = {
   title: "",
   description: "",
   startDate: "",
@@ -88,3 +77,13 @@ export const convertCalendarEventsToDialogSessions = (
       };
     });
 };
+
+export const convertDateToISODate = (date: Date) =>
+  new Date(date.getTime() - date.getTimezoneOffset() * 60000).toISOString().slice(0, 10);
+
+export const convertDateToHourMinute = (date: Date) =>
+  new Date(date.getTime() - date.getTimezoneOffset() * 60000).toISOString().slice(11, 16);
+
+export const capitalizeFirstLetter = (text: string) => (text ? text.charAt(0).toUpperCase() + text.slice(1) : text);
+
+export const convertToUtcIso = (date: string, time: string) => new Date(`${date}T${time}:00`).toISOString(); // -> "2025-08-10T07:30:00.000Z"
